requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
pydantic==2.5.0